Decouple search from tests/fixtures
